Dont use multiple change outputs in trade fee tx and blind vote
The trade protocol expects that the trade fee tx has only 1 change
output.
At testing a vote tx became invalid due a wrong fee. Not 100% clear yet
why but deactivated multiple change outputs for blind vote tx as well.

Overall that approach seems to be rather problematic and if we cannot
use it at the trade fee tx it will not help a lot either.